fix(lark-rs): widen terminal/rule priority i32 → i64 so large priorities don't tie (#352) (#384)
Priorities were parsed to i128 then clamped to i32. Two distinct priorities
both > i32::MAX saturated to the same value and tied, so lark-rs broke the tie
by name order where Python honours the true ordering (bounty H4-4).

Repro `start: A | B` / `A.5000000000: "x"` / `B.9000000000: "x"` on `x`:
Python picks B (9e9 > 5e9); lark-rs clamped both to i32::MAX, tied, picked A.
Confirmed Python accepts arbitrarily large priorities (no rejection at 9e20) —
contract is store-wide-not-reject.

Widen i32 → i64 end to end: loader clamp (Tok::Number, tokenizer), AST priority
fields, public RuleOptions/TerminalDef fields + builders, the template store
tuple, and the Earley/CYK priority accumulators (so summed large priorities
don't re-truncate). Public-API change → escalate-tier (ADR-0016/0025); Proposed
ADR-0034 added.

XFAIL h4_4_priority_i32_saturation_tie flips to passing; negative control
h4_4_priority_small_and_boundary_still_order pins ordinary/boundary ordering.

Refs #352
Part of #354

+++ lark-rs/src/grammar/loader/compiler.rs
@@ -93,7 +93,7 @@ pub(super) struct GrammarCompiler {
     /// The modifiers (`!` keep-all, `?` expand1) and priority are kept so each
     /// instantiation inherits the template's rule options, exactly as Python Lark
     /// deep-copies the template's `RuleOptions` onto every instance.
-    pub(super) templates: HashMap<String, (Vec<String>, Vec<AliasedExpansion>, String, i32)>,
+    pub(super) templates: HashMap<String, (Vec<String>, Vec<AliasedExpansion>, String, i64)>,
     /// Memo of template instantiations: canonical `name<args>` key → instance rule
     /// name. Lets a self-recursive template (`_sep{x,d}: x | _sep{x,d} d x`) resolve
     /// its own reference to the rule already being built instead of recursing
